+++ wallets/ctrl-extension/src/extension/chain-wallet.ts
@@ -1,0 +1,7 @@
+import { ChainRecord, ChainWalletBase, Wallet } from '@cosmos-kit/core';
+
+export class ChainCTRLExtension extends ChainWalletBase {
+  constructor(walletInfo: Wallet, chainInfo: ChainRecord) {
+    super(walletInfo, chainInfo);
+  }
+}