+++ wallets/ctrl-extension/src/index.ts
@@ -1,0 +1,2 @@
+export * from './ctrl';
+export * from './extension';