+++ wallets/owallet-mobile/src/wallet-connect/chain-wallet.ts
@@ -1,0 +1,9 @@
+import { ChainRecord, Wallet } from '@cosmos-kit/core';
+import { ChainWC } from '@cosmos-kit/walletconnect';
+import { OWalletClient } from './client';
+
+export class ChainOWalletMobile extends ChainWC {
+  constructor(walletInfo: Wallet, chainInfo: ChainRecord) {
+    super(walletInfo, chainInfo, OWalletClient);
+  }
+}