+++ wallets/owallet-mobile/src/wallet-connect/types.ts
@@ -1,0 +1,8 @@
+export interface OWalletAccount {
+  address: string;
+  algo: string;
+  bech32Address: string;
+  isNanoLedger: boolean;
+  name: string;
+  pubKey: string;
+}