+++ wallets/owallet/src/index.ts
@@ -1,3 +1,4 @@
 import { wallets as ext } from '@cosmos-kit/owallet-extension';
+import { wallets as mobile } from '@cosmos-kit/owallet-mobile';
 
-export const wallets = [...ext];
+export const wallets = [...ext, ...mobile];