+++ wallets/xdefi-extension/src/extension/chain-wallet.ts
@@ -1,7 +1,0 @@
-import { ChainRecord, ChainWalletBase, Wallet } from '@cosmos-kit/core';
-
-export class ChainXDEFIExtension extends ChainWalletBase {
-  constructor(walletInfo: Wallet, chainInfo: ChainRecord) {
-    super(walletInfo, chainInfo);
-  }
-}