+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.28.1](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.28.0...docs@2.28.1) (2024-12-16)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
+# [2.28.0](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.27.0...docs@2.28.0) (2024-12-16)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
+# [2.27.0](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.26.0...docs@2.27.0) (2024-12-16)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
+# [2.26.0](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.25.2...docs@2.26.0) (2024-12-16)
+
+**Note:** Version bump only for package docs
+
+
+
+
+
 ## [2.25.2](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.25.1...docs@2.25.2) (2024-10-14)
 
 **Note:** Version bump only for package docs