+++ wallets/cdcwallet/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cdcwallet@2.15.1...@cosmos-kit/cdcwallet@2.15.2) (2024-12-16)
+
+**Note:** Version bump only for package @cosmos-kit/cdcwallet
+
+
+
+
+
 ## [2.15.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cdcwallet@2.15.0...@cosmos-kit/cdcwallet@2.15.1) (2024-10-09)
 
 **Note:** Version bump only for package @cosmos-kit/cdcwallet