+++ wallets/coin98/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.13.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/coin98@2.13.1...@cosmos-kit/coin98@2.13.2) (2024-12-16)
+
+**Note:** Version bump only for package @cosmos-kit/coin98
+
+
+
+
+
 ## [2.13.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/coin98@2.13.0...@cosmos-kit/coin98@2.13.1) (2024-10-09)
 
 **Note:** Version bump only for package @cosmos-kit/coin98