+++ wallets/prax-extension/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/prax-extension@0.3.1...@cosmos-kit/prax-extension@0.3.2) (2024-12-16)
+
+**Note:** Version bump only for package @cosmos-kit/prax-extension
+
+
+
+
+
 ## [0.3.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/prax-extension@0.3.0...@cosmos-kit/prax-extension@0.3.1) (2024-10-09)
 
 **Note:** Version bump only for package @cosmos-kit/prax-extension